# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EmberRealm - Anime & Manga streaming platform API proxy
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /manga/search:
    get:
      operationId: searchManga
      tags: [manga]
      summary: Search manga by title via MangaDex
      parameters:
        - in: query
          name: q
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Manga search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
  /manga/popular:
    get:
      operationId: getPopularManga
      tags: [manga]
      summary: Get popular manga from MangaDex
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Popular manga list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
  /manga/latest:
    get:
      operationId: getLatestManga
      tags: [manga]
      summary: Get latest updated manga from MangaDex
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Latest manga list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaListResponse"
  /manga/{id}:
    get:
      operationId: getMangaDetails
      tags: [manga]
      summary: Get manga details from MangaDex
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Manga details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MangaDetails"
  /manga/{id}/chapters:
    get:
      operationId: getMangaChapters
      tags: [manga]
      summary: Get chapter list for a manga from MangaDex (English, first 500 chapters)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chapter list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterListResponse"
  /manga/chapter/{chapterId}/pages:
    get:
      operationId: getMangaChapterPages
      tags: [manga]
      summary: Get page image URLs for a chapter
      parameters:
        - in: path
          name: chapterId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Chapter pages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChapterPagesResponse"
components:
  schemas:
...
    MangaItem:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        lastChapter:
          type: ["string", "null"]
        lastUpdated:
          type: ["string", "null"]
      required:
        - id
        - title
    MangaListResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/MangaItem"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - results
        - total
        - limit
        - offset
    MangaDetails:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        altTitles:
          type: array
          items:
            type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        year:
          type: ["integer", "null"]
        tags:
          type: array
          items:
            type: string
        authors:
          type: array
          items:
            type: string
        artists:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        totalChapters:
          type: ["integer", "null"]
        lastUpdated:
          type: ["string", "null"]
      required:
        - id
        - title
    ChapterItem:
      type: object
      properties:
        id:
          type: string
        chapter:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        volume:
          type: ["string", "null"]
        publishAt:
          type: ["string", "null"]
        readableAt:
          type: ["string", "null"]
        pages:
          type: ["integer", "null"]
        translatedLanguage:
          type: string
        scanlationGroup:
          type: ["string", "null"]
      required:
        - id
        - translatedLanguage
    ChapterListResponse:
      type: object
      properties:
        chapters:
          type: array
          items:
            $ref: "#/components/schemas/ChapterItem"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
      required:
        - chapters
        - total
        - limit
        - offset
    ChapterPagesResponse:
      type: object
      properties:
        chapterId:
          type: string
        baseUrl:
          type: string
        hash:
          type: string
        pages:
          type: array
          items:
            type: string
        dataSaver:
          type: boolean
      required:
        - chapterId
        - baseUrl
        - hash
        - pages
        - dataSaver