# lib/api-zod/src/generated/types/mangaDetails.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */

export interface MangaDetails {
  id: string;
  title: string;
  altTitles?: string[];
  /** @nullable */
  description?: string | null;
  /** @nullable */
  coverUrl?: string | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  year?: number | null;
  tags?: string[];
  authors?: string[];
  artists?: string[];
  /** @nullable */
  rating?: number | null;
  /** @nullable */
  totalChapters?: number | null;
  /** @nullable */
  lastUpdated?: string | null;
}