# lib/api-zod/src/generated/types/mangaListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * EmberRealm - Anime & Manga streaming platform API proxy
 * OpenAPI spec version: 0.1.0
 */
import type { MangaItem } from './mangaItem';

export interface MangaListResponse {
  results: MangaItem[];
  total: number;
  limit: number;
  offset: number;
}